+++ app/src/tab/server_login.cpp
@@ -61,6 +61,8 @@ public:
                 .name = auth.User.Name,
                 .access_token = auth.AccessToken,
                 .server_id = auth.ServerId,
+                .is_admin = auth.User.Policy.IsAdministrator,
+                .config = std::move(auth.User.Configuration),
             };
 
             brls::sync([dialog, u, this]() {
@@ -171,7 +173,14 @@ bool ServerLogin::onSignin() {
         try {
             auto resp = HTTP::post(this->url + jellyfin::apiAuthByName, data.dump(), header);
             jellyfin::AuthResult r = nlohmann::json::parse(resp);
-            AppUser u = {.id = r.User.Id, .name = r.User.Name, .access_token = r.AccessToken, .server_id = r.ServerId};
+            AppUser u = {
+                .id = r.User.Id,
+                .name = r.User.Name,
+                .access_token = r.AccessToken,
+                .server_id = r.ServerId,
+                .is_admin = r.User.Policy.IsAdministrator,
+                .config = std::move(r.User.Configuration),
+            };
             brls::sync([ASYNC_TOKEN, u]() {
                 ASYNC_RELEASE
                 AppConfig::instance().addUser(u, this->url);